Initial commit - Film Festival Project

+++ FilmFestivalProject-main/DBMS Project/DBMS Project/src/main/java/com/filmfestival/backend/payload/response/JwtResponse.java
@@ -6,12 +6,12 @@ import lombok.Data;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private Integer id;
     private String name;
     private String email;
     private String role;
 
-    public JwtResponse(String accessToken, Long id, String name, String email, String role) {
+    public JwtResponse(String accessToken, Integer id, String name, String email, String role) {
         this.token = accessToken;
         this.id = id;
         this.name = name;
