+++ FilmFestivalProject-main/DBMS Project/DBMS Project/src/main/java/com/filmfestival/backend/repository/FilmRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmRepository extends JpaRepository<Film, Long> {
+public interface FilmRepository extends JpaRepository<Film, Integer> {
 }
