+++ FilmFestivalProject-main/DBMS Project/DBMS Project/src/main/java/com/filmfestival/backend/security/services/UserDetailsImpl.java
@@ -13,7 +13,7 @@ import java.util.Objects;
 public class UserDetailsImpl implements UserDetails {
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
     private String name;
     private String email;
 
@@ -22,7 +22,7 @@ public class UserDetailsImpl implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserDetailsImpl(Long id, String name, String email, String password,
+    public UserDetailsImpl(Integer id, String name, String email, String password,
                            Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.name = name;
@@ -47,7 +47,7 @@ public class UserDetailsImpl implements UserDetails {
         return authorities;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
